Moved logout logic to makeRequest function + Moved logout tests to independent files
