+++ sdk/interfaces/__tests__/index.test.js
@@ -1,9 +1,15 @@
-import { initialize, getToken, logout } from '../index';
+import { initialize, getToken } from '../index';
 import makeRequest from '../../requests';
 import { getParameters } from '../../configuration';
 
 jest.mock('../../requests');
 const { REQUEST_TYPES } = jest.requireActual('../../requests');
+
+const mockAddEventListener = jest.fn();
+jest.mock('react-native/Libraries/Linking/Linking', () => ({
+  addEventListener: mockAddEventListener,
+  removeEventListener: jest.fn(),
+}));
 
 describe('initialize', () => {
   const parameters = {
@@ -48,42 +54,6 @@ describe('getToken', () => {
     }
     expect(makeRequest).toHaveBeenCalledTimes(1);
     expect(makeRequest).toHaveBeenCalledWith(REQUEST_TYPES.GET_TOKEN);
-    expect.assertions(3);
-  });
-});
-
-describe('logout', () => {
-  afterEach(() => jest.clearAllMocks());
-
-  it('calls logout with correct postLogoutRedirectUri', async () => {
-    makeRequest.mockReturnValue(Promise.resolve());
-    mockAddEventListener.mockImplementation((eventType, eventHandler) => {
-      if (eventType === 'url')
-        eventHandler({
-          url: 'sdkidu.testing://redirect',
-        });
-    });
-    const redirectUri = await logout();
-    expect(makeRequest).toHaveBeenCalledTimes(1);
-    expect(makeRequest).toHaveBeenCalledWith(REQUEST_TYPES.LOGOUT);
-    expect(redirectUri).toBe('sdkidu.testing://redirect');
-  });
-
-  it('calls logout with incorrect postLogoutRedirectUri', async () => {
-    makeRequest.mockReturnValue(Promise.reject(Error()));
-    mockAddEventListener.mockImplementation((eventType, eventHandler) => {
-      if (eventType === 'url')
-        eventHandler({
-          url: '',
-        });
-    });
-    try {
-      await logout();
-    } catch (error) {
-      expect(error).toMatchObject(Error('Invalid post logout redirect uri'));
-    }
-    expect(makeRequest).toHaveBeenCalledTimes(1);
-    expect(makeRequest).toHaveBeenCalledWith(REQUEST_TYPES.LOGOUT);
     expect.assertions(3);
   });
 });
