+++ sdk/interfaces/index.js
@@ -1,4 +1,3 @@
-import { Linking } from 'react-native';
 import makeRequest, { REQUEST_TYPES } from '../requests';
 import { setParameters } from '../configuration';
 
@@ -15,32 +14,6 @@ const login = () => makeRequest(REQUEST_TYPES.LOGIN);
 
 const getToken = () => makeRequest(REQUEST_TYPES.GET_TOKEN);
 
-const logout = async () => {
-  let resolveFunction;
-  let rejectFunction;
-  const promise = new Promise((resolve, reject) => {
-    resolveFunction = resolve;
-    rejectFunction = reject;
-  });
-
-  const handleOpenUrl = event => {
-    const urlCheck = event.url;
-    if (urlCheck) {
-      setParameters({ code: 'empty' });
-      resolveFunction(urlCheck);
-    } else rejectFunction(Error('Invalid post logout redirect uri'));
-    Linking.removeEventListener('url', handleOpenUrl);
-  };
-
-  try {
-    Linking.addEventListener('url', handleOpenUrl);
-    await makeRequest(REQUEST_TYPES.LOGOUT);
-  } catch (error) {
-    Linking.removeEventListener('url', handleOpenUrl);
-    rejectFunction(Error("Couldn't make request"));
-  }
-
-  return promise;
-};
+const logout = async () => makeRequest(REQUEST_TYPES.LOGOUT);
 
 export { initialize, login, getToken, logout };
