+++ sdk/requests/__tests__/logout.test.js
@@ -1,0 +1,87 @@
+import logout from '../logout';
+import { getParameters } from '../../configuration';
+
+jest.mock('../../configuration');
+
+const mockAddEventListener = jest.fn();
+const mockLinkingOpenUrl = jest.fn(() => Promise.resolve());
+
+jest.mock('react-native/Libraries/Linking/Linking', () => ({
+  addEventListener: mockAddEventListener,
+  removeEventListener: jest.fn(),
+  openURL: mockLinkingOpenUrl,
+}));
+
+describe('logout', () => {
+  afterEach(() => jest.clearAllMocks());
+
+  it('calls logout with idTokenHint, postLogoutRedirectUri and state', async () => {
+    getParameters.mockReturnValue({
+      idToken: 'idToken',
+      postLogoutRedirectUri: 'postLogoutRedirectUri',
+      state: 'chau',
+    });
+    mockAddEventListener.mockImplementation((eventType, eventHandler) => {
+      if (eventType === 'url')
+        eventHandler({
+          url: 'postLogoutRedirectUri',
+        });
+    });
+    const redirectUri = await logout();
+    expect(mockLinkingOpenUrl).toHaveBeenCalledTimes(1);
+    expect(mockLinkingOpenUrl).toHaveBeenCalledWith(
+      'https://auth-testing.iduruguay.gub.uy/oidc/v1/logout?id_token_hint=idToken&post_logout_redirect_uri=postLogoutRedirectUri&state=chau',
+    );
+    expect(redirectUri).toBe('postLogoutRedirectUri');
+  });
+
+  it('calls logout with idTokenHint and postLogoutRedirectUri but without state', async () => {
+    getParameters.mockReturnValue({
+      idToken: 'idToken',
+      postLogoutRedirectUri: 'postLogoutRedirectUri',
+      state: '',
+    });
+    mockAddEventListener.mockImplementation((eventType, eventHandler) => {
+      if (eventType === 'url')
+        eventHandler({
+          url: 'postLogoutRedirectUri',
+        });
+    });
+    const redirectUri = await logout();
+    expect(mockLinkingOpenUrl).toHaveBeenCalledTimes(1);
+    expect(mockLinkingOpenUrl).toHaveBeenCalledWith(
+      'https://auth-testing.iduruguay.gub.uy/oidc/v1/logout?id_token_hint=idToken&post_logout_redirect_uri=postLogoutRedirectUri&state=',
+    );
+    expect(redirectUri).toBe('postLogoutRedirectUri');
+  });
+
+  it('calls logout with idTokenHint and state but without postLogoutRedirectUri', async () => {
+    getParameters.mockReturnValue({
+      idToken: 'idToken',
+      postLogoutRedirectUri: '',
+      state: 'chau',
+    });
+    try {
+      await logout();
+    } catch (error) {
+      expect(error).toMatchObject(Error("Couldn't make request"));
+    }
+    expect(mockLinkingOpenUrl).not.toHaveBeenCalled();
+    expect.assertions(1);
+  });
+
+  it('calls logout with postLogoutRedirectUri and state but without idTokenHint', async () => {
+    getParameters.mockReturnValue({
+      idToken: '',
+      postLogoutRedirectUri: 'postLogoutRedirectUri',
+      state: 'chau',
+    });
+    try {
+      await logout();
+    } catch (error) {
+      expect(error).toMatchObject(Error("Couldn't make request"));
+    }
+    expect(mockLinkingOpenUrl).not.toHaveBeenCalled();
+    expect.assertions(1);
+  });
+});
